Removed extraneous helper functions from `gauge_inference.py` to their own separate module in `inference/gauge_inference_utils.py` to improve readability and simplify the inference interface.
